AgentMCP/genesys_integration/requirements_genesys.txt
# Requirements para integração Genesys + Agent-MCP
# Dependências específicas para executar LLaMA 70B localmente

# Core LLaMA dependencies
torch>=2.0.0
torchvision
torchaudio
transformers>=4.35.0
accelerate>=0.20.0

# LLaMA C++ bindings for better performance
llama-cpp-python>=0.2.20

# Quantization and optimization
bitsandbytes>=0.41.0
optimum>=1.14.0

# Multimodal support
Pillow>=9.0.0
opencv-python>=4.8.0

# LangChain integration (from Genesys original)
langchain>=0.0.300
langchain-community>=0.0.10
langchain-core>=0.1.0

# API server
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
python-multipart>=0.0.6

# MCP integration
mcp>=1.8.1
requests>=2.31.0

# Environment and configuration
python-dotenv>=1.0.0
pydantic>=2.0.0

# Original Genesys tools dependencies
chromadb>=0.4.0
duckduckgo-search>=3.9.0
psutil>=5.9.0

# Development and debugging
pytest>=7.4.0
black>=23.0.0
isort>=5.12.0

# GPU monitoring (optional)
nvidia-ml-py3>=7.352.0
GPUtil>=1.4.0

# Fine-tuning support (Phase 2)
datasets>=2.14.0
peft>=0.5.0
trl>=0.7.0

# File handling and utilities
pathlib2>=2.3.7
typing-extensions>=4.5.0
aiofiles>=23.0.0

# Logging and monitoring
structlog>=23.1.0
rich>=13.5.0

google-generativeai